+++ src/_includes/head.njk
@@ -53,7 +53,7 @@
     crossorigin="anonymous"
   />
 
-  <link rel="stylesheet" href="{{ '/scripts/main.entry.css' | assetUrl }}" />
+  {% assetStylesheet '/assets/main.entry.css' %}
 
   <link rel="icon" href="/assets/images/favicon.png" />
