+++ src/_includes/layouts/markdown.njk
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="description" content="Christian Kaisermann - Resume" />
 
-    <link rel="stylesheet" href="{{ '/styles/markdown.entry.css' | assetUrl }}" />
+    {% assetStylesheet '/assets/styles/markdown.entry.css' %}
 
     <title>kaisermann's resume</title>
   </head>
