+++ src/assets/scripts/components/HeaderControls.svelte
