+++ src/assets/scripts/components/Remote.svelte
