+++ src/assets/scripts/components/RemoteTrigger.svelte
