+++ src/assets/scripts/components/SpaceTrigger.svelte
