+++ src/assets/scripts/components/Video.svelte
