+++ src/assets/scripts/components/Webcam.svelte
