+++ src/assets/scripts/modules/analytics.js
