+++ src/assets/scripts/modules/links.js
