+++ src/assets/scripts/modules/noise.js
