+++ src/assets/scripts/modules/textNav.js
