+++ src/assets/styles/main/partials/animations.css
